# backend/swagger.yaml
openapi: 3.0.0
info:
  title: PokeGPT API
  description: API that answers questions about Pokémon using ChatGPT.
  version: 1.0.0
servers:
  - url: 'http://localhost:3000'
paths:
  /ask:
    post:
      summary: Ask about Pokémon
      description: Sends the user's question about Pokémon to ChatGPT and receives an answer.
      operationId: askChatGPT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - question
              properties:
                question:
                  type: string
                  description: A question about Pokémon.
            examples:
              example1: 
                value:
                  question: "What is Pikachu's special move?"
              example2:
                value:
                  question: "What's your favorite Pokémon?"
              example3:
                value:
                  question: "Who is Harry Potter?"

      responses:
        '200':
          description: Successful response with the answer from ChatGPT.
          content:
            application/json:
              schema:
                type: object
                properties:
                  answer:
                    type: string
                    description: The answer generated by ChatGPT based on the Pokémon-related question.
              examples:
                example1:
                  value:
                    answer: "Pikachu's special move is called Volt Tackle."
                example2:
                  value:
                    answer: "As an AI, I do not have personal preferences."
                example3:
                  value:
                    answer: "I'm sorry, I can only answer questions about Pokémon."
        '400':
          description: Bad Request - The question field is missing or invalid.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
              examples:
                example1:
                  value:
                    answer: "Please enter a valid question."
                example2:
                  value:
                    answer: "Please enter a meaningful question, not a random string"
        '500':
          description: Internal Server Error - There was an error processing the question.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "There was an error processing your question"
components:
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: Authorization
      description: "You must provide an OpenAI API key in the Authorization header. Format: 'Bearer YOUR_API_KEY'"
  security:
    - apiKey: []